refactor: changing user cart page route
